requirements.txt
python-telegram-bot
requests
pymongo
apscheduler
Pillow
python-decouple